Accept pointer to int as TcpRouteMapping's TTL
- change TTL from uint16 to int

[#119118969]

+++ cmd/router-configurer/main.go
@@ -164,7 +164,7 @@ func main() {
 	logger.Debug("creating-routing-api-client", lager.Data{"api-location": routingAPIAddress})
 	routingAPIClient := routing_api.NewClient(routingAPIAddress)
 
-	updater := routing_table.NewUpdater(logger, &routingTable, configurer, routingAPIClient, uaaClient, clock, uint16(defaultRouteExpiry.Seconds()))
+	updater := routing_table.NewUpdater(logger, &routingTable, configurer, routingAPIClient, uaaClient, clock, int(defaultRouteExpiry.Seconds()))
 
 	ticker := clock.NewTicker(*staleRouteCheckInterval)
